# api/api.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Flight Booking System
  description: A series of APIs for Flight Booking System
servers:
  - url: http://localhost:8080
paths:
  /liveness:
    get:
      responses:
        "200":
          description: liveness response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pong"

  /api/v1/flights/search:
    get:
      summary: Search flights with filtering, sorting, and pagination
      description: Returns a list of flights based on search criteria with pagination support
      operationId: searchFlights
      parameters:
        - name: departure_date
          in: query
          schema:
            type: string
            format: date
          description: Date of departure (YYYY-MM-DD)
          example: "2025-01-20"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [departure_time, arrival_time, base_price, available_seats]
            default: departure_time
          description: Field to sort the results by
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order (ascending or descending)
        - name: filters
          in: query
          style: deepObject
          explode: true
          schema:
            type: object
            additionalProperties:
              type: string
          description: |
            Key-value pairs for filtering records. Available filters:
            - departure_city: Departure city name
            - arrival_city: Arrival city name
            - airline: Airline name
            - flight_number: Flight number

            Example: filters[departure_city]=New York&filters[arrival_city]=London&filters[airline]=British Airways
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchFlightResponse"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/v1/orders:
    post:
      summary: Submit a new flight booking order
      description: Creates a new order for flight booking
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Order"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Pong:
      type: object
      required:
        - startTime
      properties:
        startTime:
          type: string

    Flight:
      type: object
      required:
        - id
        - flight_number
        - airline
        - departure_city
        - arrival_city
        - departure_time
        - arrival_time
        - aircraft
        - status
        - available_seats
        - base_price
      properties:
        id:
          type: integer
          format: uint
          example: 1
        flight_number:
          type: string
          example: "BA123"
          minLength: 1
          maxLength: 20
        airline:
          type: string
          example: "British Airways"
          minLength: 1
          maxLength: 100
        departure_city:
          type: string
          example: "New York"
          minLength: 1
          maxLength: 100
        arrival_city:
          type: string
          example: "London"
          minLength: 1
          maxLength: 100
        departure_time:
          type: string
          format: date-time
          example: "2025-01-20T10:00:00Z"
        arrival_time:
          type: string
          format: date-time
          example: "2025-01-20T22:00:00Z"
        aircraft:
          type: string
          example: "Boeing 787"
          minLength: 1
          maxLength: 50
        status:
          type: string
          enum: [SCHEDULED, DELAYED, CANCELLED, IN_PROGRESS, COMPLETED]
          example: "SCHEDULED"
        available_seats:
          type: integer
          example: 150
          minimum: 0
        base_price:
          type: integer
          description: Price in smallest currency unit (e.g., cents)
          example: 50000
          minimum: 0

    SearchFlightResponse:
      type: object
      required:
        - data
        - totalCount
        - page
        - pageSize
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Flight"
        totalCount:
          type: integer
          format: int64
          minimum: 0
          description: Total number of records
        page:
          type: integer
          minimum: 1
          description: Current page number
        pageSize:
          type: integer
          minimum: 1
          description: Number of items per page

    CreateOrderRequest:
      type: object
      required:
        - flight_id
        - customer_id
        - ticket_amount
      properties:
        flight_id:
          type: integer
          format: uint
          example: 1
          description: ID of the flight to book
        customer_id:
          type: integer
          format: uint
          example: 1
          description: ID of the customer making the booking
        ticket_amount:
          type: integer
          minimum: 1
          example: 2
          description: Number of tickets to book

    Order:
      type: object
      required:
        - id
        - flight_id
        - customer_id
        - status
        - total_amount
        - order_number
        - booking_time
      properties:
        id:
          type: integer
          format: uint
          example: 1
        flight_id:
          type: integer
          format: uint
          example: 1
        customer_id:
          type: integer
          format: uint
          example: 1
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
          example: "PENDING"
        total_amount:
          type: integer
          description: Total amount in smallest currency unit (e.g., cents)
          example: 50000
        order_number:
          type: string
          example: "ORD123456789"
        booking_time:
          type: string
          format: date-time
          example: "2025-01-20T10:00:00Z"
        flight:
          $ref: "#/components/schemas/Flight"
        customer:
          $ref: "#/components/schemas/Customer"

    Customer:
      type: object
      required:
        - name
        - email
        - phone
      properties:
        id:
          type: integer
          format: uint
          example: 1
          readOnly: true
        name:
          type: string
          example: "John Doe"
          minLength: 1
          maxLength: 100
        email:
          type: string
          example: "john.doe@example.com"
          format: email
          minLength: 1
          maxLength: 100
        phone:
          type: string
          example: "0912345678"
          minLength: 1
          maxLength: 20

    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Error code
        message:
          type: string
          description: Error message